Remove keyboard layout from VM description
This is a follow-up of previous commit.

Now that we allow user to configure the keyboard configuration, we can't
hardcode a value of "US keyboard layout" in the VM description.

Moreover, forwarding this setting to the VM description is clunky, since
the keyboard layout is set in the first build stage (rootfs), but the VM
description is written in the postinst step that comes after the image
is created.

To keep things simple, just drop it from the description.